Adding unit test for content endpoint
